requirements-lock.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
cycler==0.12.1
Cython==3.2.4
easydict==1.13
filelock==3.25.2
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
onnx==1.21.0
onnxruntime-gpu==1.21.0
opencv-contrib-python-headless==4.10.0.84
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==12.1.1
prettytable==3.17.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.16
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.7
